server/requirements.txt
anyio==3.6.2
black==22.12.0
charset-normalizer==3.0.1
click==8.1.3
cohere==3.1.6
fastapi==0.89.1
flake8==6.0.0
flake8-black==0.3.6
h11==0.14.0
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.1
pandas==1.5.3
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
